+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 #### Added
 
+- Report all parsing errors in Markdown files (#389, @NathanReb)
+
 #### Changed
 
 #### Deprecated
@@ -11,6 +13,7 @@
 - Fixed compatibility with Cmdliner 1.1.0 (#371, @Leonidas-from-XIV)
 - Report errors and exit codes of toplevel directives (#382, @talex5,
   @Leonidas-from-XIV)
+- Fix block locations in error reporting (#389, @NathanReb)
 
 #### Removed